requirements.txt
albucore==0.0.24
albumentations==2.0.8
alembic==1.18.4
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cloudpickle==3.1.2
contourpy==1.3.3
cryptography==46.0.5
cycler==0.12.1
databricks-sdk==0.100.0
docker==7.1.0
fastapi==0.135.1
filelock==3.25.2
Flask==3.1.3
flask-cors==6.0.2
fonttools==4.62.1
fsspec==2026.2.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
graphene==3.4.3
graphql-core==3.2.8
graphql-relay==3.2.0
gunicorn==25.1.0
h11==0.16.0
huey==2.6.0
idna==3.11
importlib_metadata==8.7.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.5.0
llvmlite==0.46.0
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
mlflow==3.10.1
mlflow-skinny==3.10.1
mlflow-tracing==3.10.1
mpmath==1.3.0
narwhals==2.18.0
networkx==3.6.1
numba==0.64.0
numpy==2.4.3
opencv-python==4.13.0.92
opencv-python-headless==4.13.0.92
opentelemetry-api==1.40.0
opentelemetry-proto==1.40.0
opentelemetry-sdk==1.40.0
opentelemetry-semantic-conventions==0.61b0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
plotly==6.6.0
prettytable==3.17.0
protobuf==6.33.5
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pynndescent==0.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scikit-learn==1.8.0
scipy==1.17.1
seaborn==0.13.2
setuptools==82.0.1
simsimd==6.5.16
six==1.17.0
skops==0.13.0
smmap==5.0.3
SQLAlchemy==2.0.48
sqlparse==0.5.5
starlette==0.52.1
streamlit==1.55.0
stringzilla==4.6.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
toml==0.10.2
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tornado==6.5.5
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
urllib3==2.6.3
uvicorn==0.42.0
wcwidth==0.6.0
Werkzeug==3.1.6
zipp==3.23.0
